Organize GCS storage methods
