+++ pbm/storage/gcs/google_client.go
@@ -4,19 +4,14 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
-	"path"
 	"strings"
-	"time"
 
 	storagegcs "cloud.google.com/go/storage"
 	"github.com/googleapis/gax-go/v2"
 	"golang.org/x/oauth2/google"
-	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
 
 	"github.com/percona/percona-backup-mongodb/pbm/errors"
-	"github.com/percona/percona-backup-mongodb/pbm/storage"
 )
 
 func (g *GCS) initClient() error {
@@ -128,192 +123,4 @@ func shouldRetryExtended(err error) bool {
 	}
 
 	return false
-}
-
-func (g *GCS) save(name string, data io.Reader, options ...storage.Option) error {
-	opts := storage.GetDefaultOpts()
-	for _, opt := range options {
-		if err := opt(opts); err != nil {
-			return errors.Wrap(err, "processing options for save")
-		}
-	}
-
-	ctx := context.Background()
-	w := g.bucketHandle.Object(path.Join(g.cfg.Prefix, name)).NewWriter(ctx)
-	if g.cfg.parallelUploadEnabled() {
-		if g.log != nil && opts.UseLogger {
-			g.log.Debug(`uploading %q [size hint: %v (%v); parallel upload part size: %v (%v); concurrency: %d]`,
-				name,
-				opts.Size, storage.PrettySize(opts.Size),
-				g.cfg.ChunkSize, storage.PrettySize(int64(g.cfg.ChunkSize)),
-				g.cfg.ParallelUploadConcurrency)
-		}
-
-		w.EnableParallelUpload = true
-		w.ParallelUploadConfig = storagegcs.ParallelUploadConfig{
-			PartSize:       g.cfg.ChunkSize,
-			MaxConcurrency: g.cfg.ParallelUploadConcurrency,
-		}
-	} else {
-		const align int64 = 256 << 10 // 256 KiB (both min size and alignment)
-
-		partSize := storage.ComputePartSize(
-			opts.Size,
-			defaultChunkSize,
-			align,
-			10_000,
-			int64(g.cfg.ChunkSize),
-		)
-
-		if rem := partSize % align; rem != 0 {
-			partSize += align - rem
-		}
-
-		if g.log != nil && opts.UseLogger {
-			g.log.Debug(`uploading %q [size hint: %v (%v); part size: %v (%v)]`,
-				name,
-				opts.Size, storage.PrettySize(opts.Size),
-				partSize, storage.PrettySize(partSize))
-		}
-
-		w.ChunkSize = int(partSize)
-		w.ChunkRetryDeadline = g.cfg.Retryer.ChunkRetryDeadline
-	}
-	if g.log != nil && opts.UseLogger {
-		w.ProgressFunc = func(written int64) {
-			if opts.Size > 0 {
-				g.log.Debug("uploaded %v / %v (%.1f%%)",
-					written, opts.Size,
-					float64(written)*100/float64(opts.Size))
-			} else {
-				g.log.Debug("uploaded %v (total unknown)", written)
-			}
-		}
-	}
-
-	if _, err := io.Copy(w, data); err != nil {
-		return errors.Wrap(err, "save data")
-	}
-
-	if err := w.Close(); err != nil {
-		return errors.Wrap(err, "writer close")
-	}
-
-	return nil
-}
-
-func (g *GCS) fileStat(name string) (storage.FileInfo, error) {
-	ctx := context.Background()
-
-	attrs, err := g.bucketHandle.Object(path.Join(g.cfg.Prefix, name)).Attrs(ctx)
-	if err != nil {
-		if errors.Is(err, storagegcs.ErrObjectNotExist) {
-			return storage.FileInfo{}, storage.ErrNotExist
-		}
-
-		return storage.FileInfo{}, errors.Wrap(err, "get properties")
-	}
-
-	inf := storage.FileInfo{
-		Name: attrs.Name,
-		Size: attrs.Size,
-	}
-
-	if inf.Size == 0 {
-		return inf, storage.ErrEmpty
-	}
-
-	return inf, nil
-}
-
-func (g *GCS) list(prefix, suffix string) ([]storage.FileInfo, error) {
-	ctx := context.Background()
-
-	var files []storage.FileInfo
-	it := g.bucketHandle.Objects(ctx, &storagegcs.Query{Prefix: prefix})
-
-	for {
-		attrs, err := it.Next()
-
-		if errors.Is(err, iterator.Done) {
-			break
-		}
-
-		if err != nil {
-			return nil, errors.Wrap(err, "list objects")
-		}
-
-		name := attrs.Name
-		name = strings.TrimPrefix(name, prefix)
-		if len(name) == 0 {
-			continue
-		}
-		if name[0] == '/' {
-			name = name[1:]
-		}
-
-		if suffix != "" && !strings.HasSuffix(name, suffix) {
-			continue
-		}
-
-		files = append(files, storage.FileInfo{
-			Name: name,
-			Size: attrs.Size,
-		})
-	}
-
-	return files, nil
-}
-
-func (g *GCS) delete(name string) error {
-	ctx := context.Background()
-
-	err := g.bucketHandle.Object(path.Join(g.cfg.Prefix, name)).Delete(ctx)
-	if err != nil {
-		if errors.Is(err, storagegcs.ErrObjectNotExist) {
-			return storage.ErrNotExist
-		}
-		return errors.Wrap(err, "delete object")
-	}
-
-	return nil
-}
-
-func (g *GCS) copy(src, dst string) error {
-	ctx := context.Background()
-
-	srcObj := g.bucketHandle.Object(path.Join(g.cfg.Prefix, src))
-	dstObj := g.bucketHandle.Object(path.Join(g.cfg.Prefix, dst))
-
-	_, err := g.fileStat(src)
-	if err == storage.ErrNotExist {
-		return err
-	}
-
-	_, err = dstObj.CopierFrom(srcObj).Run(ctx)
-	return err
-}
-
-func (g *GCS) getPartialObject(name string, buf *storage.Arena, start, length int64) (io.ReadCloser, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*60)
-	defer cancel()
-
-	obj := g.bucketHandle.Object(path.Join(g.cfg.Prefix, name))
-	reader, err := obj.NewRangeReader(ctx, start, length)
-	if err != nil {
-		if errors.Is(err, storagegcs.ErrObjectNotExist) || isRangeNotSatisfiable(err) {
-			return nil, io.EOF
-		}
-
-		return nil, storage.GetObjError{Err: err}
-	}
-
-	ch := buf.GetSpan()
-	_, err = io.CopyBuffer(ch, reader, buf.CpBuf)
-	if err != nil {
-		ch.Close()
-		return nil, errors.Wrap(err, "copy")
-	}
-	reader.Close()
-	return ch, nil
 }
